PLF-7090 : upgrade to Tomcat 8.5
