+++ pom.xml
@@ -40,13 +40,16 @@
     <!-- Servers -->
     <version.jboss.as>7.1.1.Final</version.jboss.as>
     <version.jetty8>8.1.7.v20120910</version.jetty8>
-    <version.tomcat7>7.0.30</version.tomcat7>
+    <version.tomcat8>8.5.13</version.tomcat8>
 
     <!-- For testing only -->
-    <version.arquillian>1.0.0.Final</version.arquillian>
-    <version.arquillian.tomcat>1.0.0.CR4</version.arquillian.tomcat>
-    <version.shrinkwrap>1.0.0-cr-1</version.shrinkwrap>
-    <version.shrinkwrap.resolver>2.0.0-alpha-1</version.shrinkwrap.resolver>
+    <version.arquillian>1.1.12.Final</version.arquillian>
+    <version.arquillian.tomcat>1.0.0</version.arquillian.tomcat>
+    <version.arquillian.glassfish>1.0.0.CR3</version.arquillian.glassfish>
+    <version.arquillian.jetty>1.0.0.CR3</version.arquillian.jetty>
+    <version.shrinkwrap>1.2.6</version.shrinkwrap>
+    <version.shrinkwrap.resolver>2.2.5</version.shrinkwrap.resolver>
+    <version.servlet-api>3.1.0</version.servlet-api>
   </properties>
 
   <dependencyManagement>
@@ -78,7 +81,7 @@
       </dependency>
       <dependency>
         <groupId>org.exoplatform.gatein.wci</groupId>
-        <artifactId>wci-tomcat7</artifactId>
+        <artifactId>wci-tomcat8</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
@@ -183,22 +186,22 @@
       <dependency>
         <groupId>org.apache.tomcat</groupId>
         <artifactId>tomcat-catalina</artifactId>
-        <version>${version.tomcat7}</version>
+        <version>${version.tomcat8}</version>
       </dependency>
       <dependency>
         <groupId>org.apache.tomcat.embed</groupId>
         <artifactId>tomcat-embed-core</artifactId>
-        <version>${version.tomcat7}</version>
+        <version>${version.tomcat8}</version>
       </dependency>
       <dependency>
         <groupId>org.apache.tomcat.embed</groupId>
         <artifactId>tomcat-embed-jasper</artifactId>
-        <version>${version.tomcat7}</version>
+        <version>${version.tomcat8}</version>
       </dependency>
       <dependency>
-        <groupId>org.apache.tomcat.embed</groupId>
-        <artifactId>tomcat-embed-logging-juli</artifactId>
-        <version>${version.tomcat7}</version>
+        <groupId>javax.servlet</groupId>
+        <artifactId>javax.servlet-api</artifactId>
+        <version>${version.servlet-api}</version>
       </dependency>
 
       <!-- JBoss 7 -->
