+++ test/servers/pom.xml
@@ -11,7 +11,7 @@
 
    <modules>
      <module>glassfish3</module>
-     <module>tomcat7</module>
+     <module>tomcat8</module>
      <!--<module>jboss7</module>-->
      <module>jetty8</module>
    </modules>
