--- test/servers/tomcat7/pom.xml
@@ -1,62 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-   <parent>
-      <groupId>org.exoplatform.gatein.wci</groupId>
-      <artifactId>wci-test-servers</artifactId>
-      <version>5.0.x-SNAPSHOT</version>
-   </parent>
-   <modelVersion>4.0.0</modelVersion>
-   <artifactId>wci-test-tomcat7</artifactId>
-   <packaging>jar</packaging>
-   <name>GateIn - WCI test Tomcat7</name>
-
-  <dependencies>
-
-    <dependency>
-      <groupId>org.exoplatform.gatein.wci</groupId>
-      <artifactId>wci-test-core</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.exoplatform.gatein.wci</groupId>
-      <artifactId>wci-test-core</artifactId>
-      <type>test-jar</type>
-    </dependency>
-
-    <!-- -->
-    <dependency>
-      <groupId>org.exoplatform.gatein.wci</groupId>
-      <artifactId>wci-tomcat7</artifactId>
-    </dependency>
-
-    <!-- Arquillian container -->
-    <dependency>
-      <groupId>org.jboss.arquillian.container</groupId>
-      <artifactId>arquillian-tomcat-embedded-7</artifactId>
-      <version>${version.arquillian.tomcat}</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.tomcat.embed</groupId>
-      <artifactId>tomcat-embed-core</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.tomcat.embed</groupId>
-      <artifactId>tomcat-embed-jasper</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.tomcat.embed</groupId>
-      <artifactId>tomcat-embed-logging-juli</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.jdt.core.compiler</groupId>
-      <artifactId>ecj</artifactId>
-      <version>3.7</version>
-    </dependency>
-    <dependency>
-      <groupId>javax.servlet.jsp</groupId>
-      <artifactId>jsp-api</artifactId>
-      <version>2.2</version>
-    </dependency>
-
-  </dependencies>
-
-</project>
