--- test/servers/tomcat7/src/test/resources/org/gatein/wci/test/tomcat7/authentication/context.xml
@@ -1,5 +1,0 @@
-<Context debug="0" reloadable="true" privileged="true">
-
-  <Realm className="org.gatein.wci.test.tomcat7.AuthenticationRealm"/>
-
-</Context>
