--- test/servers/tomcat7/src/test/resources/org/gatein/wci/test/tomcat7/deployment/context.xml
@@ -1,3 +1,0 @@
-<Context privileged="true">
-
-</Context>
