--- test/servers/tomcat7/src/test/resources/org/gatein/wci/test/tomcat7/requestdispatch/context.xml
@@ -1,3 +1,0 @@
-<Context privileged="true">
-
-</Context>
