+++ test/servers/tomcat8/pom.xml
@@ -1,4 +1,3 @@
-
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
    <parent>
       <groupId>org.exoplatform.gatein.wci</groupId>
@@ -6,15 +5,12 @@
       <version>5.0.x-SNAPSHOT</version>
    </parent>
    <modelVersion>4.0.0</modelVersion>
-   <artifactId>wci-test-glassfish3</artifactId>
+   <artifactId>wci-test-tomcat8</artifactId>
    <packaging>jar</packaging>
-   <name>GateIn - WCI test GlassFish 3</name>
+   <name>GateIn - WCI test Tomcat8</name>
 
   <dependencies>
-    <dependency>
-      <groupId>org.exoplatform.gatein.wci</groupId>
-      <artifactId>wci-glassfish3</artifactId>
-    </dependency>
+
     <dependency>
       <groupId>org.exoplatform.gatein.wci</groupId>
       <artifactId>wci-test-core</artifactId>
@@ -25,18 +21,38 @@
       <type>test-jar</type>
     </dependency>
 
+    <!-- -->
+    <dependency>
+      <groupId>org.exoplatform.gatein.wci</groupId>
+      <artifactId>wci-tomcat8</artifactId>
+    </dependency>
+
     <!-- Arquillian container -->
     <dependency>
       <groupId>org.jboss.arquillian.container</groupId>
-      <artifactId>arquillian-glassfish-embedded-3.1</artifactId>
-      <version>${version.arquillian.glassfish}</version>
+      <artifactId>arquillian-tomcat-embedded-8</artifactId>
+      <version>${version.arquillian.tomcat}</version>
       <scope>test</scope>
     </dependency>
     <dependency>
-      <groupId>org.glassfish.extras</groupId>
-      <artifactId>glassfish-embedded-web</artifactId>
-      <version>3.1</version>
-      <scope>test</scope>
+      <groupId>org.apache.tomcat.embed</groupId>
+      <artifactId>tomcat-embed-core</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.apache.tomcat.embed</groupId>
+      <artifactId>tomcat-embed-jasper</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.eclipse.jdt.core.compiler</groupId>
+      <artifactId>ecj</artifactId>
+      <version>4.4.2</version>
+    </dependency>
+    <dependency>
+      <groupId>javax.servlet.jsp</groupId>
+      <artifactId>jsp-api</artifactId>
+      <version>2.2</version>
+    </dependency>
+
   </dependencies>
+
 </project>
