+++ test/servers/tomcat8/src/test/resources/org/gatein/wci/test/tomcat8/crosscontext/context.xml
@@ -1,0 +1,6 @@
+<Context debug="0" reloadable="true" privileged="true">
+
+  <Realm className="org.gatein.wci.test.tomcat8.AuthenticationRealm"/>
+  <Loader delegate="true"/>
+
+</Context>
