+++ test/servers/tomcat8/src/test/resources/org/gatein/wci/test/tomcat8/requestdispatch/context.xml
@@ -1,0 +1,3 @@
+<Context reloadable="true" privileged="true">
+    <Loader delegate="true"/>
+</Context>
