+++ test/servers/tomcat8/src/test/resources/org/gatein/wci/test/tomcat8/servletlistener/web.xml
