--- tomcat/tomcat7/pom.xml
@@ -1,23 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-   <parent>
-      <groupId>org.exoplatform.gatein.wci</groupId>
-      <artifactId>wci-tomcat</artifactId>
-      <version>5.0.x-SNAPSHOT</version>
-   </parent>
-   <modelVersion>4.0.0</modelVersion>
-   <artifactId>wci-tomcat7</artifactId>
-   <packaging>jar</packaging>
-   <name>GateIn - WCI Tomcat 7 compatibility component</name>
-
-   <dependencies>
-      <dependency>
-         <groupId>org.exoplatform.gatein.wci</groupId>
-         <artifactId>wci-wci</artifactId>
-      </dependency>
-      <dependency>
-         <groupId>org.apache.tomcat</groupId>
-         <artifactId>tomcat-catalina</artifactId>
-         <scope>provided</scope>
-      </dependency>
-   </dependencies>
-</project>
