+++ tomcat/tomcat8/src/main/java/org/gatein/wci/command/TomcatCommandDispatcher.java
