+++ tomcat/tomcat8/src/main/java/org/gatein/wci/tomcat/TC7WebAppContext.java
